Cambiado tamaño
git-svn-id: http://svn.smartick.es/svn/sva-smartick/trunk/smartickAndroid@6101 79f48f48-82ef-0310-9b43-d5d80899f613

+++ src/com/smartick/ScreenUtils.java
@@ -18,7 +18,7 @@ public class ScreenUtils {
         } catch (NoSuchMethodError e) { 
         	width = display.getWidth();
         }
-        Double val = new Double(width)/new Double(1000);
+        Double val = new Double(width)/new Double(1200);
         val = val * 100d;
         return val.intValue();
     }
